Move `NamespaceName` from `libsql-wal` to `libsql-sys`
