+++ libsql-wal/src/lib.rs
@@ -3,7 +3,6 @@
 pub mod bottomless;
 pub mod error;
 pub mod io;
-pub mod name;
 pub mod registry;
 pub mod replication;
 pub mod segment;
